+++ dbc_influxdb/__init__.py
@@ -1,1 +1,4 @@
+from .logger import setup_logging
 from .main import dbcInflux
+
+__all__ = ["dbcInflux", "setup_logging"]